refactor: 과제별로 파일 분리, 레벨2 하다 말음
